third_party/WebKit/LayoutTests/external/wpt/fetch/api/response/response-static-redirect-expected.txt

This is a testharness.js-based test.
FAIL Check default redirect response assert_equals: expected "" but got "OK"
FAIL Check response returned by static method redirect(), status = 301 assert_equals: expected "" but got "OK"
FAIL Check response returned by static method redirect(), status = 302 assert_equals: expected "" but got "OK"
FAIL Check response returned by static method redirect(), status = 303 assert_equals: expected "" but got "OK"
FAIL Check response returned by static method redirect(), status = 307 assert_equals: expected "" but got "OK"
FAIL Check response returned by static method redirect(), status = 308 assert_equals: expected "" but got "OK"
PASS Check error returned when giving invalid url to redirect()
PASS Check error returned when giving invalid status to redirect(), status = 200
PASS Check error returned when giving invalid status to redirect(), status = 309
PASS Check error returned when giving invalid status to redirect(), status = 400
PASS Check error returned when giving invalid status to redirect(), status = 500
Harness: the test ran to completion.
